Fix repo change

Approved-by: Dmitri Babaev <dmitri.lb@gmail.com>

+++ dltranz/train.py
@@ -21,7 +21,7 @@ def batch_to_device(batch, device, non_blocking):
     x, y = batch
     if not isinstance(x, dict):
         new_x = {k: v.to(device=device, non_blocking=non_blocking) if isinstance(v, torch.Tensor) else v for k, v in x.payload.items()}
-        new_y = y.to(device=device, non_blocking=non_blocking).float()
+        new_y = y.to(device=device, non_blocking=non_blocking)
         return PaddedBatch(new_x, x.seq_lens), new_y
     else:
         batches = {}
@@ -29,7 +29,7 @@ def batch_to_device(batch, device, non_blocking):
             new_x = {k: v.to(device=device, non_blocking=non_blocking) if isinstance(v, torch.Tensor) else v for k, v in
                      sx.payload.items()}
             batches[key] = PaddedBatch(new_x, sx.seq_lens)
-        new_y = y.to(device=device, non_blocking=non_blocking).float()
+        new_y = y.to(device=device, non_blocking=non_blocking)
         return batches, new_y
